Move assert to its own util
