+++ src/cli/generate/DevDB.ts
@@ -1,12 +1,12 @@
 import * as fsp from 'node:fs/promises'
 import {Config} from 'alinea/core/Config'
 import type {UploadResponse} from 'alinea/core/Connection'
-import {createId} from 'alinea/core/Id'
-import {getWorkspace} from 'alinea/core/Internal'
 import {type CommitRequest, checkCommit} from 'alinea/core/db/CommitRequest'
 import {LocalDB} from 'alinea/core/db/LocalDB'
+import {createId} from 'alinea/core/Id'
+import {getWorkspace} from 'alinea/core/Internal'
 import {CachedFSSource} from 'alinea/core/source/FSSource'
-import {assert} from 'alinea/core/source/Utils'
+import {assert} from 'alinea/core/util/Assert'
 import {keys, values} from 'alinea/core/util/Objects'
 import {
   basename,
