+++ src/core/db/EntryIndex.ts
@@ -23,7 +23,8 @@ import {hashBlob} from '../source/GitUtils.js'
 import {ShaMismatchError} from '../source/ShaMismatchError.js'
 import {bundleContents, type Source} from '../source/Source.js'
 import {ReadonlyTree} from '../source/Tree.js'
-import {assert, compareStrings} from '../source/Utils.js'
+import {compareStrings} from '../source/Utils.js'
+import {assert} from '../util/Assert.js'
 import {sourceChanges} from './CommitRequest.js'
 import {EntryTransaction} from './EntryTransaction.js'
 import {IndexEvent} from './IndexEvent.js'
