+++ src/core/db/EntryResolver.ts
@@ -29,7 +29,8 @@ import {hasExact} from 'alinea/core/util/Checks'
 import {entries, fromEntries} from 'alinea/core/util/Objects'
 import {unreachable} from 'alinea/core/util/Types'
 import * as cito from 'cito'
-import {assert, compareStrings} from '../source/Utils.js'
+import {compareStrings} from '../source/Utils.js'
+import {assert} from '../util/Assert.js'
 import type {EntryFilter, EntryIndex} from './EntryIndex.js'
 import {LinkResolver} from './LinkResolver.js'
