+++ src/core/source/Utils.ts
@@ -1,12 +1,5 @@
 import {crypto} from '@alinea/iso'
 
-export function assert(value: unknown, message?: string): asserts value {
-  if (value) return
-  const error = new Error(message)
-  if ('captureStackTrace' in Error) Error.captureStackTrace(error, assert)
-  throw error
-}
-
 /**
  * Computes the SHA-1 hash of the input data.
  */
